Started to switch from NSLogger → PonyDebugger
Successfully logging from PonyDebugger:

- added PonyDebugger submodule
- found out that you need to enableCoreDataDebugging to get expandable objects
- turned off CBL LogTo stuff

As an aside, also started work on an AppleScript app (Pony.app) that should hopefully make getting to a line in code directly from the logs doable via a specially registered pony:// URL handler.

Also got rid of build number increment pre-build script... need to figure out if that makes a difference or not for making logging via Pony easier or not?